% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.topo.path.compo.R
\name{get.topo.path.compo}
\alias{get.topo.path.compo}
\title{Topological path composition}
\usage{
get.topo.path.compo(df, nnetwork = net, orig = p)
}
\arguments{
\item{df}{a data.frame containing a pto column of destination patches. Names in df$pto should match vertex names in nnetwork}

\item{nnetwork}{a network of connection from the igraph package}

\item{orig}{name of the origin patch, should match a vertex name in nnetwork}
}
\value{
A data.table containing origin-destination patches (from, to), the length of the shortest paths (length) and the number of paths (num.path)
connecting the origin-destination patches.
}
\description{
get.topo.path.compo calculates all topological shortest paths between an origin patch and destinations patch(es) and returns
the number of paths, their length and their composition (i.e., intermediate patches along the path).
}
