% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paraNorm.R
\name{paraNorm}
\alias{paraNorm}
\title{Matrix normalization}
\usage{
paraNorm(mymat)
}
\arguments{
\item{mymat}{the matrix to normalize}
}
\value{
a normalized matrix
}
\description{
paraNorm normalizes a distance matrix
}
