% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLC-SCP-package.R
\docType{package}
\name{RFLC-SCP-package}
\alias{RFLC-SCP}
\alias{RFLC-SCP-package}
\title{RFLC-SCP: A Robust Framework to assess Landscape Connectivity for Systematic Conservation Planning application}
\description{
This project provides all the data and codes used to produce the methodological framework and results presented in: Prima, M.-C., Renaud, J., Witté, I., Suarez, L., Rouveyrol, P., Fernando, M., Sacchi, A., Cosentino, F., Santini, L., Maiorano, L., Moreira, F., Dertien, J., & Thuiller, W. (in prep.). A robust framework to assess landscape connectivity for systematic conservation planning application. The workflow is as follow, Step 1. Clustering species in functional groups having similar traits and environmental niches (analyses/Rcode01_GetFunctionalGroups.R). Step 2. Generating resistance and habitat suitability maps for each group based on an ensemble of species distribution models (analyses/Rcode2_GetResisSuitMaps.R). Step 3. Generating ecological continuities per group (analyses/Rcode3_GetEcologicalContinuities.R). Step 4. Calculating multi-scale network metrics (analyses/Rcode4_GetConnectivityMetrics.R). Step 5. (optional) Plot figures (analyses/Rcode5_GetFigures.R). First run the make.R file before any run of R scripts. R functions are located in the R/ folder and are automatically loaded when the make.R file is run. Function help can be found by running the classic linecode help('function_name'). Initial and intermediate datasets are located is the data/ folder (to be downloaded and added to the project, see the Readme file) Generated outputs are located in the outputs/ folder (to be downloaded and added to the project, see the Readme file).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcpri3/RFLC-SCP}
  \item Report bugs at \url{https://github.com/mcpri3/RFLC-SCP/issues}
}

}
\author{
\strong{Maintainer}: Marie-Caroline Prima \email{marie-caroline.prima@univ-grenoble-alpes.fr} (\href{https://orcid.org/{{0000-0001-9392-0022}}}{ORCID}) [copyright holder]

}
\keyword{internal}
