% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesClustering.R
\name{speciesClustering}
\alias{speciesClustering}
\title{Create clusters based on dissimilarity matrix}
\usage{
speciesClustering(mat.species.DIST, opt.no_clust_max = 15)
}
\arguments{
\item{mat.species.DIST}{a list containing a matrix object corresponding to the
dissimilarity distance between each pair of species.}

\item{opt.no_clust_max}{an integer
corresponding to the maximum number of clusters to be tested.}
}
\value{
A \code{list} containing:

\describe{
\item{clust.dendrograms}{a \code{list} with an object of
class \code{\link[stats]{hclust}}}
\item{clust.evaluation}{ \cr
\describe{
\item{\code{no.clusters}}{number of clusters used for the clustering}
\item{\code{variable}}{evaluation metrics' name}
\item{\code{value}}{value of evaluation metric \cr \cr}
}
}
\item{plot.clustNo}{\code{ggplot2} object, representing the different
values of metrics to choose the number of clusters \cr \cr}
}
}
\description{
speciesClustering creates clusters of species based
on a distance matrix between those species. Several metrics are computed
to evaluate these clusters and a graphic is produced to help the user to
choose the best number of clusters.
}
\details{
This function allows to \strong{obtain dendrograms based on a dissimilarity
distance matrix between species}.
The process is as follows :

\describe{
\item{\strong{1. Choice of the \cr optimal \cr clustering method}}{
hierarchical clustering on the dissimilarity matrix is realized with the
\code{\link[stats]{hclust}}.
\itemize{
\item Several methods are available for the agglomeration :
\emph{complete}, \emph{ward.D}, \emph{ward.D2}, \emph{single},
\emph{average (UPGMA)}, \emph{mcquitty (WPGMA)}, \emph{median (WPGMC)}
and \emph{centroid (UPGMC)}.
\item \emph{Mouchet et al. (2008)} proposed a similarity measure between
the input distance and the one obtained with the clustering which must
be minimized to help finding the best clustering method :
\deqn{ 1 - cor( \text{mat.species.DIST}, \text{clustering.DIST} ) ^ 2}
}
\strong{For each agglomeration method, this measure is calculated. The
method that minimizes it is kept and used for further analyses. \cr \cr}
}

\item{\strong{2. Evaluation of the \cr clustering}}{once the hierarchical
clustering is done, the number of clusters to keep should be chosen. \cr
To do that, several metrics are computed :
\itemize{
\item{\emph{Dunn index (\code{mdunn}) : }}{ratio of the smallest
distance between observations not in the same cluster to the largest
intra-cluster distance. Value between \code{0} and \eqn{\infty}, and
should be maximized.}
\item{\emph{Meila's Variation of Information index (\code{mVI}) : }}
{measures the amount of information lost and gained in changing
between two clusterings. Should be minimized.}
\item{\emph{Coefficient of determination (\code{R2}) : }}{value
between \code{0} and \code{1}. Should be maximized.}
\item{\emph{Calinski and Harabasz index (\code{ch}) : }}{the higher
the value, the "better" is the solution.}
\item{\emph{Corrected rand index (\code{Rand}) : }}{measures the
similarity between two data clusterings. Value between \code{0} and
\code{1}, with \code{0} indicating that the two data clusters do not
agree on any pair of points and \code{1} indicating that the data
clusters are exactly the same.}
\item{\emph{Average silhouette width (\code{av.sil}) : }}{Observations
with a large \code{s(i)} (almost \code{1}) are very well clustered, a
small \code{s(i)} (around \code{0}) means that the observation lies
between two clusters, and observations with a negative \code{s(i)} are
probably placed in the wrong cluster. Should be maximized.}
}
\strong{A graphic is produced, giving the values of these metrics in
function of the number of clusters used. Number of clusters are
highlighted in function of evaluation metrics' values to help the
user to make his/her optimal choice : the brighter (yellow-ish) the
better.}
}
}

\emph{\cr \cr
Mouchet M., Guilhaumon f., Villeger S., Mason N.W.H., Tomasini J.A. &
Mouillot D., 2008. Towards a consensus for calculating dendrogam-based
functional diversity indices. Oikos, 117, 794-800.}
}
