% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesDistanceOverlap.R
\name{speciesDistanceOverlap}
\alias{speciesDistanceOverlap}
\title{Computation of niche overlap distances between species}
\usage{
speciesDistanceOverlap(mat.overlap.object)
}
\arguments{
\item{mat.overlap.object}{a list with tab.PA (a \code{matrix} or \code{data.frame} with
sites in rows and species in columns, containing either \code{NA},
\code{0} or \code{1}) and tab.env (a \code{matrix} or \code{data.frame} with
sites in rows and environmental variables in columns)}
}
\value{
A \code{matrix} containing overlap distances between each pair
of species, calculated as \code{1 - Schoeners D}.
}
\description{
speciesDistanceOverlap generates a distance matrix between
species, based on co-occurrence of species.
}
